+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.2.0 (unreleased)
 
+- Fixed error with `dup` and `clone`
 - Dropped support for Ruby < 3
 
 ## 0.1.8 (2023-02-01)